+++ source/Program.php
@@ -18,7 +18,7 @@ namespace AsocialMedia\Sfera;
  * 
  * @author     ASOCIAL MEDIA Maciej Strączkowski <biuro@asocial.media>
  * @copyright  ASOCIAL MEDIA Maciej Strączkowski <biuro@asocial.media>
- * @version    2.0.1
+ * @version    2.1.0
  */
 class Program
 {