+++ CHANGELOG.md
@@ -1,12 +1,20 @@
 # CHANGELOG
 
+## 2.1.0 (2021-01-14)
+
+- Added AsocialMedia\Sfera\Navireo for Navireo support (iqa)
+- Updated README
+- Updated version number to 2.1.0
+- Updated year in LICENSE
+- Updated description in composer.json
+
 ## 2.0.1 (2020-03-06)
 
 - Changed RuntimeException message if ext-com_dotnet is not available
 - Added missing phpdoc @param tags
 - Fixed composer / package conflict
 - Updated version number to 2.0.1
-- Changed year in LICENSE
+- Updated year in LICENSE
 
 ## 2.0.0 (2019-10-12)